Added Lighting To Scene Interactable Objects
There is now global lighting on all prefabs, this is just changing the
SpriteRenderer to use a diffuse material rather then a default one. This
allowed for the character to have a light bulb above his head when he
was near something he could interact with. Like the Icing Dispenser
which changes the icing the character is using.